+++ examples/merge_taxa.R
@@ -1,5 +1,5 @@
 ## Merge Cyperus papyrus and Cyperus dives
 summary(Easplist, c(206, 197))
 
-sp_list <- merge_taxa(object = Easplist, concepts = c(206, 197),
+merged_cyperus <- merge_taxa(object = Easplist, concepts = c(206, 197),
     print_output = TRUE)